chat-app-backend/backend/chats/serializers.py
from rest_framework import serializers

from .models import User, Message

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('name',)

class MessageSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Message
        fields = ('author', 'content', 'creation_time')
