chat-app-backend/requirements.txt
Django==3.0.4
psycopg2==2.8.4
djangorestframework==3.11.0
